Move generated enum content to an enums.go file
